wip: extract hash password
