+++ src/kairos/shared/security/password/index.ts
@@ -1,3 +1,2 @@
-export { hashPassword } from './password-composition';
 export { passwordHasher } from './password-hasher';
 export type { PasswordHasher } from './password-hasher';
